requirements.txt
requests
tabulate
colorama
pyyaml
